+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-04-25
+
+### Changed
+
+- Upgrade to Ruthless 0.7.0, Gradle 7.6.1, Java 11
+- Transform `GradleVersion` utility into `CurrentGradleVersion` to avoid confusion with `org.gradle.util.GradleVersion`. Also removed unused methods.
+
+## [0.3.0] - 2022-12-17
+
+### Changed
+
+- Upgrade to Gradle 7.5.1
+
 ## [0.2.1] - 2022-07-17
 
 ### Changed
@@ -38,7 +51,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `GradleVersion`
 - `VersionComparator`
 
-[Unreleased]: https://github.com/LajosCseppento/gradle-plugin-common/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/LajosCseppento/gradle-plugin-common/compare/v0.4.0...HEAD
+
+[0.3.0]: https://github.com/LajosCseppento/gradle-plugin-common/releases/tag/v0.4.0
+
+[0.3.0]: https://github.com/LajosCseppento/gradle-plugin-common/releases/tag/v0.3.0
 
 [0.2.1]: https://github.com/LajosCseppento/gradle-plugin-common/releases/tag/v0.2.1